# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is API documentation on DataBx Bromo2 Microservices",
    "version": "1.0.0",
    "title": "DataBx Bromo2 Microservices",
    "contact": {
      "email": "akbar.muhammadakbarmaulana@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags": [
    {
      "name": "file",
      "description": "Simple interact with file mongodb collection"
    }
  ],
  "paths": {
    "/select/file": {
      "get": {
        "tags": [
          "file"
        ],
        "summary": "Get file by given document id",
        "description": "This should be return only one file ya",
        "operationId": "RetrieveFile",
        "security": [
          {"basicAuth":  []}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/xClientId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "200": {
            "description": "RetrieveFile success, with given request",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Check request, header not comply with request rules",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "401": {
            "description": "You are not permitted!"
          },
          "500": {
            "description": "General error, system cannot tell you what wrongs",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/insert/file": {
      "post": {
        "tags": [
          "file"
        ],
        "summary": "Store document on file collection",
        "description": "Return only document profile",
        "operationId": "StoreFile",
        "security": [
          {"basicAuth":  []}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/xClientId"
          },
          {
            "$ref": "#/components/parameters/Filename"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/File"
        },
        "responses": {
          "200": {
            "description": "StoreFile success, with given request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Check request, body not comply with request rules",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "401": {
            "description": "You are not permitted!"
          },
          "500": {
            "description": "General error, system cannot tell you what wrongs",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/delete/file": {
      "delete": {
        "tags": [
          "file"
        ],
        "summary": "Delete file based on document id",
        "description": "Return if state whether success or not",
        "operationId": "RemoveFile",
        "security": [
          {"basicAuth":  []}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/xClientId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "200": {
            "description": "RemoveFile success, with given request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Check request, header not comply with request rules",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "401": {
            "description": "You are not permitted!"
          },
          "500": {
            "description": "General error, system cannot tell you what wrongs",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:7720/bromo2/api/v1"
    }
  ],
  "components": {
    "requestBodies": {
      "File": {
        "required": true,
        "content": {
          "application/octet-stream": {
            "schema": {
              "type": "string",
              "format": "binary"
            }
          }
        }
      }
    },
    "schemas": {
      "RequestFile": {
        "type": "string",
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "type": "string",
            "maxLength": 36,
            "pattern": "[A-Za-z0-9]{8}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{12}"
          },
          "data": {
            "type": "object",
            "properties": {
              "document_id": {
                "type": "string"
              },
              "filename" : {
                "type": "string"
              }
            }
          }
        }
      },
      "Response": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "response_id": {
            "type": "string",
            "maxLength": 36,
            "pattern": "[A-Za-z0-9]{8}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{12}"
          },
          "request_id": {
            "type": "string",
            "maxLength": 36,
            "pattern": "[A-Za-z0-9]{8}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{12}"
          },
          "success": {
            "type": "boolean"
          },
          "response_time": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[0-9]$"
          },
          "message": {
            "type": "string",
            "maxLength": 256,
            "pattern": "^.$"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "parameters": {
      "xClientId": {
        "name": "X-Client-ID",
        "in": "header",
        "description": "Custom Id that defines you as caller",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 32,
          "pattern": "[A-Za-z0-9]{32}"
        }
      },
      "Filename": {
        "name": "filename",
        "in": "header",
        "description": "how do you want to name it in documents, either with extension or not",
        "required": true,
        "schema": {
          "type": "string",
          "maxLength": 128
        }
      },
      "RequestId": {
        "name": "request_id",
        "in": "header",
        "description": "Request id that accompanies, what else request",
        "required": true,
        "schema": {
          "type": "string",
          "maxLength": 36,
          "pattern": "[A-Za-z0-9]{8}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{12}"
        }
      },
      "DocumentId": {
        "name": "document_id",
        "in": "header",
        "description": "id that generated by mongodb to store documents",
        "required": true,
        "schema": {
          "type": "string",
          "maxLength": 24,
          "pattern": "[A-Za-z0-9]{24}"
        }
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}